pyproject.toml
[tool.poetry]
name = "bigbiro"
version = "0.1.0"
description = "Esoteric language (Transpiles to C++)"
authors = ["Meetesh Saini <code.meeteshsaini@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
ply = "^3.11"
click = "^8.1.7"
pyyaml = "^6.0.1"
requests = "^2.31.0"


[tool.poetry.group.dev.dependencies]
black = "^23.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 80